requirements.txt
pandas>=2.0
sqlalchemy>=2.0
psycopg2-binary
pyarrow          # for parquet
openpyxl         # for excel
lxml             # for read_html
python-dateutil